Migrate HttpServerSpec into EndToEndTestSpec
