+++ subproject/presentation/build.gradle.kts
@@ -13,10 +13,4 @@ dependencies {
     implementation(KTOR_SERIALIZATION)
 
     implementation(TYPESAFE_CONFIG)
-
-    testImplementation(testFixtures(LOGGING))
-
-    testFixturesImplementation(KTOR_CLIENT_CORE)
-    testFixturesImplementation(KTOR_CLIENT_CIO)
-    testFixturesImplementation(KTOR_CLIENT_SERIALIZATION)
 }
